+++ migrations/20260211193000_next_up_default_off/migration.sql
@@ -1,0 +1,6 @@
+-- AlterTable
+ALTER TABLE "User" ALTER COLUMN "nextUpEnabled" SET DEFAULT false;
+
+-- Data migration
+UPDATE "User" SET "nextUpEnabled" = false WHERE "nextUpEnabled" = true;
+